[C] Refactor min and tagged flow control so that min is a special case of tagged.
